fix auth

+++ routes/web.php
@@ -18,9 +18,9 @@ use App\Http\Controllers\QrCodeController;
 |
 */
 
-Route::get('/Admin/Dashboard',[App\Http\Controllers\DashboardController::class, 'index'])->name('admin.dashboard');
+Route::get('/Admin/Dashboard',[App\Http\Controllers\DashboardController::class, 'index'])->name('admin.dashboard')->middleware('staff.auth');
 
-Route::prefix('Admin/Movie')->name('admin.')->middleware('auth:staff')->group(function () {
+Route::prefix('Admin/Movie')->name('admin.')->middleware('staff.auth')->group(function () {
 
     Route::get('/', [\App\Http\Controllers\MovieController::class, 'index'])->name('movies.index');
     Route::get('/create', [\App\Http\Controllers\MovieController::class, 'create'])->name('movies.create');
@@ -29,7 +29,7 @@ Route::prefix('Admin/Movie')->name('admin.')->middleware('auth:staff')->group(fu
     Route::put('/{movie}/edit', [\App\Http\Controllers\MovieController::class, 'update'])->name('movies.update');
     Route::delete('/{movie}/delete', [\App\Http\Controllers\MovieController::class, 'destroy'])->name('movies.destroy');
 });
-Route::prefix('Admin/Category')->name('admin.')->middleware('auth:staff')->group(function () {
+Route::prefix('Admin/Category')->name('admin.')->middleware('staff.auth')->group(function () {
     Route::get('/', [\App\Http\Controllers\CategoryController::class, 'index'])->name('categories.index');
     Route::get('/create', [\App\Http\Controllers\CategoryController::class, 'create'])->name('categories.create');
     Route::post('/create', [\App\Http\Controllers\CategoryController::class, 'store'])->name('categories.store');
@@ -38,7 +38,7 @@ Route::prefix('Admin/Category')->name('admin.')->middleware('auth:staff')->group
     Route::delete('/{category}/delete', [\App\Http\Controllers\CategoryController::class, 'destroy'])->name('categories.destroy');
 });
 
-Route::prefix('Admin/Actor')->name('admin.')->middleware('auth:staff')->group(function () {
+Route::prefix('Admin/Actor')->name('admin.')->middleware('staff.auth')->group(function () {
     Route::get('/', [\App\Http\Controllers\ActorController::class, 'index'])->name('actors.index');
     Route::get('/create', [\App\Http\Controllers\ActorController::class, 'create'])->name('actors.create');
     Route::post('/create', [\App\Http\Controllers\ActorController::class, 'store'])->name('actors.store');
@@ -47,7 +47,7 @@ Route::prefix('Admin/Actor')->name('admin.')->middleware('auth:staff')->group(fu
     Route::delete('/{actor}/delete', [\App\Http\Controllers\ActorController::class, 'destroy'])->name('actors.destroy');
 });
 
-Route::prefix('Admin/Director')->name('admin.')->middleware('auth:staff')->group(function () {
+Route::prefix('Admin/Director')->name('admin.')->middleware('staff.auth')->group(function () {
     Route::get('/', [\App\Http\Controllers\DirectorController::class, 'index'])->name('directors.index');
     Route::get('/create', [\App\Http\Controllers\DirectorController::class, 'create'])->name('directors.create');
     Route::post('/create', [\App\Http\Controllers\DirectorController::class, 'store'])->name('directors.store');
@@ -65,7 +65,7 @@ Route::prefix('Login')->name('login.')->group(function(){
 
 });
 
-Route::prefix('Admin/Customer')->name('admin.')->middleware('auth:staff')->group(function(){
+Route::prefix('Admin/Customer')->name('admin.')->middleware('staff.auth')->group(function(){
 
     Route::get('/' , [App\Http\Controllers\CustomerController::class, 'index'])->name('customers.index');
     Route::delete('/{customer}/delete' , [App\Http\Controllers\CustomerController::class, 'destroy'])->name('customers.destroy');
@@ -73,19 +73,19 @@ Route::prefix('Admin/Customer')->name('admin.')->middleware('auth:staff')->group
     Route::post('/{customer}/information' , [App\Http\Controllers\CustomerController::class, 'resetPassword'])->name('customers.resetPassword');
 });
 
-Route::prefix('Admin/Staff')->name('admin.')->middleware('auth:staff')->group(function(){
-    Route::get('/', [App\Http\Controllers\StaffController::class, 'index'])->name('staffs.index');
-    Route::get('/create', [App\Http\Controllers\StaffController::class, 'create'])->name('staffs.create');
-    Route::post('/create', [App\Http\Controllers\StaffController::class, 'register'])->name('staffs.store');
-    Route::get('/{staff}/edit', [\App\Http\Controllers\StaffController::class, 'edit'])->name('staffs.edit');
-    Route::put('/{staff}/edit', [\App\Http\Controllers\StaffController::class, 'update'])->name('staffs.update');
-    Route::delete('/{staff}/delete', [\App\Http\Controllers\StaffController::class, 'destroy'])->name('staffs.destroy');
+Route::prefix('Admin/Staff')->name('admin.')->group(function(){
+    Route::get('/', [App\Http\Controllers\StaffController::class, 'index'])->name('staffs.index')->middleware('staff.auth');
+    Route::get('/create', [App\Http\Controllers\StaffController::class, 'create'])->name('staffs.create')->middleware('staff.auth');
+    Route::post('/create', [App\Http\Controllers\StaffController::class, 'register'])->name('staffs.store')->middleware('staff.auth');
+    Route::get('/{staff}/edit', [\App\Http\Controllers\StaffController::class, 'edit'])->name('staffs.edit')->middleware('staff.auth');
+    Route::put('/{staff}/edit', [\App\Http\Controllers\StaffController::class, 'update'])->name('staffs.update')->middleware('staff.auth');
+    Route::delete('/{staff}/delete', [\App\Http\Controllers\StaffController::class, 'destroy'])->name('staffs.destroy')->middleware('staff.auth');
     Route::get('/login', [\App\Http\Controllers\StaffController::class, 'showLoginForm'])->name('staffs.login');
     Route::post('/login', [\App\Http\Controllers\StaffController::class, 'login'])->name('staffs.login.check_login');
     Route::get('/logout', [\App\Http\Controllers\StaffController::class, 'logout'])->name('staffs.logout');
 });
 
-Route::prefix('Admin/Room')->name('admin.')->group(function(){
+Route::prefix('Admin/Room')->name('admin.')->middleware('staff.auth')->group(function(){
     Route::get('/', [App\Http\Controllers\RoomController::class, 'index'])->name('rooms.index');
     Route::get('/create', [App\Http\Controllers\RoomController::class, 'create'])->name('rooms.create');
     Route::post('/create', [App\Http\Controllers\RoomController::class, 'store'])->name('rooms.store');
@@ -94,7 +94,7 @@ Route::prefix('Admin/Room')->name('admin.')->group(function(){
     Route::delete('/{room}/delete', [\App\Http\Controllers\RoomController::class, 'destroy'])->name('rooms.destroy');
 });
 
-Route::prefix('Admin/Schedule') -> name('admin.') -> group(function(){
+Route::prefix('Admin/Schedule') -> name('admin.')->middleware('staff.auth')-> group(function(){
     Route::get('/', [App\Http\Controllers\ScheduleController::class, 'index'])->name('schedules.index');
     Route::get('/create', [App\Http\Controllers\ScheduleController::class, 'create'])->name('schedules.create');
     Route::post('/create', [App\Http\Controllers\ScheduleController::class, 'store'])->name('schedules.store');
@@ -109,14 +109,14 @@ Route::prefix('/')->group(function(){
     Route::get('/{movie}/detail', [App\Http\Controllers\MovieController::class, 'detail'])->name('detail');
     Route::get('/{movie_actor}/actor', [App\Http\Controllers\ActorController::class, 'show'])->name('actor');
     Route::get('/{movie_director}/director', [App\Http\Controllers\DirectorController::class, 'show'])->name('director');
-    Route::get('/{user}/user', [App\Http\Controllers\CustomerController::class, 'show'])->name('user')->middleware('auth:customers');
-    Route::put('/{user}/user', [\App\Http\Controllers\CustomerController::class, 'update'])->name('user.update')->middleware('auth:customers');
-    Route::post('/{user}/user', [\App\Http\Controllers\CustomerController::class, 'changeAvt'])->name('user.changeAvt')->middleware('auth:customers');
-    Route::get('/{schedule}/order', [App\Http\Controllers\ScheduleController::class, 'showSchedule'])->name('order')->middleware('auth:customers');
-    Route::put('/{seat_id}/{schedule_id}/order', [\App\Http\Controllers\ScheduleController::class, 'orderTicket'])->name('orderTicket')->middleware('auth:customers');
-    Route::put('/{schedule_id}/book', [\App\Http\Controllers\ScheduleController::class, 'bookTicket'])->name('bookTicket')->middleware('auth:customers');
-    Route::post('/{schedule_id}/book', [App\Http\Controllers\ScheduleController::class, 'undonScheduleBook'])->name('undon')->middleware('auth:customers');
+    Route::get('/{user}/user', [App\Http\Controllers\CustomerController::class, 'show'])->name('user')->middleware('customer.auth');
+    Route::put('/{user}/user', [\App\Http\Controllers\CustomerController::class, 'update'])->name('user.update')->middleware('customer.auth');
+    Route::post('/{user}/user', [\App\Http\Controllers\CustomerController::class, 'changeAvt'])->name('user.changeAvt')->middleware('customer.auth');
+    Route::get('/{schedule}/order', [App\Http\Controllers\ScheduleController::class, 'showSchedule'])->name('order')->middleware('customer.auth');
+    Route::put('/{seat_id}/{schedule_id}/order', [\App\Http\Controllers\ScheduleController::class, 'orderTicket'])->name('orderTicket')->middleware('customer.auth');
+    Route::put('/{schedule_id}/book', [\App\Http\Controllers\ScheduleController::class, 'bookTicket'])->name('bookTicket')->middleware('customer.auth');
+    Route::post('/{schedule_id}/book', [App\Http\Controllers\ScheduleController::class, 'undonScheduleBook'])->name('undon')->middleware('customer.auth');
     Route::get('/search', [App\Http\Controllers\MovieController::class, 'search'])->name('search');
 });
 
-Route::get('/{schedule}/{user}/generate-qrcode', [QrCodeController::class, 'index'])->name('qrcode')->middleware('auth:customers');
+Route::get('/{schedule}/{user}/generate-qrcode', [QrCodeController::class, 'index'])->name('qrcode')->middleware('customer.auth');
